Update - Fix Timesheet, TDedicacion set as Integer

+++ Bovis.Common.Model/Tables/TB_Timesheet_Proyecto.cs
@@ -10,7 +10,7 @@ namespace Bovis.Common.Model.Tables
         [Column("NukidProyecto"), NotNull] public int IdProyecto { get; set; }
         [Column("Chdescripcion"), NotNull] public string Descripcion { get; set; }
         [Column("Numdias"), NotNull] public float Dias { get; set; }
-        [Column("NumTdedicacion"), NotNull] public float TDedicacion { get; set; }
+        [Column("NumTdedicacion"), NotNull] public int TDedicacion { get; set; }
         [Column("Numcosto"), NotNull] public decimal Costo { get; set; }
         [Column("Boactivo"), NotNull] public bool Activo { get; set; }
     }
